+++ app/i18n/i18next.d.ts
@@ -1,0 +1,9 @@
+import type resources from "~/locales";
+
+declare module "i18next" {
+  interface CustomTypeOptions {
+    defaultNS: "translation";
+    resources: (typeof resources)["en"];
+    returnNull: false;
+  }
+}